+++ src/main/java/org/springframework/samples/petclinic/owner/OwnerRepository.java
@@ -57,6 +57,6 @@ public interface OwnerRepository extends Repository<Owner, Integer> {
 	 * @throws org.springframework.dao.DataRetrievalFailureException if not found
 	 */	
 	@Query("SELECT owner FROM Owner owner left join fetch owner.pets WHERE owner.id =:id")
-	public Owner findById(@Param("id") int id);
+	public Owner findById(@Param("id") Long id);
 
 }
